Create variable.php
This file I use for two different versions of the Add to Cart button inside variable products for WooCommerce. I'm looking to find a way to make this easier to show two instances of the same button.

One view on line 28 is for Desktop. That is the standard display for a single product page.
The second is on line 85 and that is for mobile only. From there I use swatches in a swipe manner. 

This method has increased conversions tenfold making the user experience much better. However, I wouldn't call two instances of the same snippet. I need to redo it to show a conditional use of Desktop and or Mobile display this way, etc.